merge object-pools-recover-transaction-queries into feature/object-pools

See merge request liquiduniverse/deepreader!16

+++ DeepReader.Types/FlattenedTypes/FlattenedBlock.cs
@@ -30,7 +30,7 @@ public class FlattenedBlock
         var obj = new FlattenedBlock
         {
             Id = reader.ReadBytes(32),
-            Number = reader.ReadUInt16(),
+            Number = reader.ReadUInt32(),
             Producer = reader.ReadName(),
             ProducerSignature = reader.ReadBytes(64),
         };
